+++ changelog.md
@@ -7,7 +7,7 @@
 - Rendering non-numeric sensor states e.g. binary sensors (#63) - (@maxwroc)
 
 ### FIXED
-- Rendering initial cached state - (@maxwroc)
+- Rendering initial cached state (#117) - (@maxwroc)
 
 ## v0.7.0