+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### ADDED
 - Grouping by date (#78) - (@maxwroc)
+- Grouping by hour
 - Point calculation / aggregation functions (#78) - (@maxwroc)
 - Smoothing setting - (@maxwroc)
 - Rendering non-numeric sensor states e.g. binary sensors (#63) - (@maxwroc)